fixed some bugs and added decayed uniformMutationPerc

+++ benchmarks/benchmark_flatsurface.cpp
@@ -1,22 +1,25 @@
 #include "benchmarks.h"
 
-class FlatSurfaceFitness : public FitnessFunction {
+class FlatSurfaceFitness : public FitnessFunction
+{
 public:
     FlatSurfaceFitness() {}
 
-    float eval(const RealChromosome &g) {
+    float eval(const RealChromosome &g)
+    {
         float fitness = 0.0;
-        for(int i=0; i<g.gene.size(); i++) {
+        for (size_t i = 0; i < g.gene.size(); i++)
+        {
             fitness += fabs(floor(g.gene[i]));
         }
         return fitness;
     }
 };
 
-
-void benchmark_flatSurface(RealGAOptions opt, GAResults &results) {
-    vector<float> LB = { -5.12, -5.12, -5.12, -5.12, -5.12 };
-    vector<float> UB = { 5.12,  5.12,  5.12,  5.12,  5.12};
+void benchmark_flatSurface(RealGAOptions opt, GAResults &results)
+{
+    vector<float> LB = {-5.12, -5.12, -5.12, -5.12, -5.12};
+    vector<float> UB = {5.12, 5.12, 5.12, 5.12, 5.12};
 
     FlatSurfaceFitness *myFitnessFunction = new FlatSurfaceFitness();
     results.name = "Flat surface";
