+++ benchmarks/benchmark_shpere.cpp
@@ -1,22 +1,26 @@
 #include "benchmarks.h"
 
-class SphereFitness : public FitnessFunction {
+class SphereFitness : public FitnessFunction
+{
 public:
     SphereFitness() {}
 
-    float eval(const RealChromosome &g) {
+    float eval(const RealChromosome &g)
+    {
         float sse = 0.0;
-        for(int i=0; i<g.gene.size(); i++) {
-            float dx = g.gene[i]/5.12;
-            sse += dx*dx;
+        for (size_t i = 0; i < g.gene.size(); i++)
+        {
+            float dx = g.gene[i] / 5.12;
+            sse += dx * dx;
         }
         return sse;
     }
 };
 
-void benchmark_sphere(RealGAOptions opt, GAResults &results) {
-    vector<float> LB = { -5.12, -5.12, -5.12, -5.12 };
-    vector<float> UB = { 5.12,  5.12,  5.12,  5.12};
+void benchmark_sphere(RealGAOptions opt, GAResults &results)
+{
+    vector<float> LB = {-5.12, -5.12, -5.12, -5.12};
+    vector<float> UB = {5.12, 5.12, 5.12, 5.12};
 
     SphereFitness *myFitnessFunction = new SphereFitness();
 
@@ -30,7 +34,7 @@ void benchmark_sphere(RealGAOptions opt, GAResults &results) {
 
     RealGA ga;
     ga.init(opt, myFitnessFunction, false);
-    
+
     benchmarkRealGen(ga, results.maxIter, 1e-4, results, BENCH_NUM_EXPERIMENTS);
 
     delete myFitnessFunction;
