+++ src/crossover_uniform.cpp
@@ -1,19 +1,24 @@
 #include "crossover.h"
 
-
-UniformCrossover::UniformCrossover(size_t chromosomeSize) {
+UniformCrossover::UniformCrossover(size_t chromosomeSize)
+{
     mChromosomeSize = chromosomeSize;
 }
 
-UniformCrossover::~UniformCrossover() {
-
+UniformCrossover::~UniformCrossover()
+{
 }
 
-void UniformCrossover::crossover(RealChromosome &a, RealChromosome &b, RealChromosome &offspring) {
-    for(int j=0; j<mChromosomeSize; j++) {
-        if(Stat::randUniform()<0.5) {
+void UniformCrossover::crossover(RealChromosome &a, RealChromosome &b, RealChromosome &offspring)
+{
+    for (size_t j = 0; j < mChromosomeSize; j++)
+    {
+        if (Stat::randUniform() < 0.5)
+        {
             offspring.gene[j] = a.gene[j];
-        } else {
+        }
+        else
+        {
             offspring.gene[j] = b.gene[j];
         }
     }
