+++ tests/testcommon.cpp
@@ -1,7 +1,9 @@
 #include "testcommon.h"
 
-void coutColor(const char * s, TextColor color) {
-    switch(color) {
+void coutColor(const char *s, TextColor color)
+{
+    switch (color)
+    {
     case TEXT_RED:
         cout << "\033[31m" << s << "\033[0m";
         break;
@@ -15,31 +17,36 @@ void coutColor(const char * s, TextColor color) {
     cout << endl;
 }
 
-
-void printDistribution(vector<float> &x, int bins, int nstars) {
+void printDistribution(vector<float> &x, int bins, int nstars)
+{
     int *freq = new int[bins];
-    for (int i=0; i<bins; ++i) freq[i] = 0;
-    
-    if(freq == NULL) {
+    for (int i = 0; i < bins; ++i)
+        freq[i] = 0;
+
+    if (freq == NULL)
+    {
         cerr << "Error! memory not allocated." << endl;
         exit(-1);
     }
-    
+
     float minV = *std::min_element(x.begin(), x.end());
     float maxV = *std::max_element(x.begin(), x.end());
     float delta = maxV - minV;
-    
-    for (int i=0; i<x.size(); ++i) {
-        int value = (int)(bins*(x[i] - minV) / delta);
-        if(value == bins) value--;
+
+    for (size_t i = 0; i < x.size(); ++i)
+    {
+        int value = (int)(bins * (x[i] - minV) / delta);
+        if (value == bins)
+            value--;
         ++freq[value];
     }
-    
-    for (int i=0; i<bins; ++i) {
+
+    for (int i = 0; i < bins; ++i)
+    {
         cout << ": ";
-        cout << string(freq[i] * nstars/x.size(), '*') << endl;
+        cout << string(freq[i] * nstars / x.size(), '*') << endl;
     }
     cout << endl;
-    
+
     delete[] freq;
 }
